bluez5_utils.mk: install automatically btgatt-server
As starting with blueZ 5.29 the time profile is no longer supported using
bluetoothd we include the btgatt-server by default in the rootfs.

btgatt-server is automatically built however, it is important to note that
buildroot does not install this example to the target by default. To install it
automatically, the .mk file for blueZ in the buildroot system is modified.

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>