kconfig: allow for mounted dl and output folder
Kconfig uses C's rename() function, which is not able to e.g. build inside
a docker container with a mounted 'dl' and 'output' folder from outside
the docker container, so to speak move over "cross-device".

The approach of the patch is a separate file_move() implementation which does
not have the issue and tries to keep the impact as small as possible.